Move leftmost indices calculations inside piecewise_poly
